Move GDPR component to src/components/
